refactor(components): removed unused experimental components
