+++ docs/cloud/connectivity/aws-connectivity.mdx
@@ -21,7 +21,7 @@ keywords:
   - term
 ---
 
-import { ToolTipTerm, DiscoverableDisclosure, CaptionedImage, JsonTable } from '@site/src/components';
+import { DiscoverableDisclosure, CaptionedImage, JsonTable } from '@site/src/components';
 
 <style>{`
   .docusaurus-mermaid-container .edgeLabel,
