+++ docs/cloud/connectivity/gcp-connectivity.mdx
@@ -21,7 +21,7 @@ keywords:
   - term
 ---
 
-import { ToolTipTerm, DiscoverableDisclosure, JsonTable } from '@site/src/components';
+import { DiscoverableDisclosure, JsonTable } from '@site/src/components';
 
 [Google Cloud Private Service Connect](https://cloud.google.com/vpc/docs/private-service-connect) allows you to open a path to Temporal without opening a public egress.
 It establishes a private connection between your Google Virtual Private Cloud (VPC) and Temporal Cloud.
