+++ docs/cloud/connectivity/index.mdx
@@ -18,7 +18,7 @@ keywords:
   - term
 ---
 
-import { ToolTipTerm, DiscoverableDisclosure, CaptionedImage, SdkTabs } from '@site/src/components';
+import { DiscoverableDisclosure, CaptionedImage, SdkTabs } from '@site/src/components';
 import { LANGUAGE_TAB_GROUP, getLanguageLabel } from '@site/src/constants/languageTabs';
 
 ## Private network connectivity for namespaces
