+++ docs/cloud/high-availability/enable.mdx
@@ -6,8 +6,6 @@ slug: /cloud/high-availability/enable
 description: Add a replica to a Namespace to enable High Availability, then manage forwarding behavior and automatic failover settings.
 ---
 
-import { ToolTipTerm } from '@site/src/components';
-
 You can enable [High Availability](/cloud/high-availability) features for a new or existing Namespace by adding a
 replica. When you add a replica, Temporal Cloud begins asynchronously replicating ongoing and existing Workflow
 Executions.
