+++ docs/cloud/rto-rpo.mdx
@@ -20,8 +20,6 @@ tags:
   - Temporal Cloud
 ---
 
-import { ToolTipTerm } from '@site/src/components';
-
 When a cloud outage disrupts a Namespace, Temporal Cloud takes measures to maintain the Namespace's availability and
 data durability. The time it takes to recover from the outage is called the _recovery time_. The _recovery point_ is how
 far back in time data must be recovered from after an outage. A durable system should have a low recovery time and a
