+++ src/components/index.js
@@ -20,7 +20,7 @@ export { default as ZoomingImage } from './images/ZoomingImage';
 
 // Information components
 export { default as DiscoverableDisclosure } from './info/DiscoverableDisclosure';
-export { default as ToolTipTerm } from './info/ToolTipTerm';
+export { default as ToolTipTerm } from './ToolTipTerm/ToolTipTerm';
 export { RelatedReadContainer, RelatedReadItem } from './info/RelatedRead/RelatedRead';
 export { default as RelatedReadList } from './info/RelatedRead/RelatedReadList';
 export { default as ReleaseNoteHeader } from './info/ReleaseNoteHeader/ReleaseNoteHeader';
